daima/requirements.txt
Flask==2.0.1
supabase==1.0.3
pandas==1.3.3
yfinance==0.1.70
pytz==2021.1
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1 